build 17 - misc

+++ paintball2/client/ref.h
@@ -299,7 +299,7 @@ typedef struct
 	qboolean	(*Vid_GetModeInfo) (int *width, int *height, int mode);
 //jitmenu	void		(*Vid_MenuInit)( void );
 	void		(*Vid_NewWindow) (int width, int height);
-	void		*(*Z_Malloc) (int size); // jitmalloc
+	void		*(*Z_Malloc) (size_t size); // jitmalloc
 	void		(*Z_Free) (void *ptr); // jitmalloc
 	qboolean	(*M_MenuActive) (void); // jitmenu, jitlinux
 	void		(*M_MouseMove) (int, int); // jitmenu, jitlinux
